requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
attrs==22.2.0
    # via
    #   jsonschema
    #   pytest
black==22.12.0
    # via -r requirements.in
cachetools==5.2.0
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via black
exceptiongroup==1.1.0
    # via pytest
google-api-core==2.11.0
    # via google-api-python-client
google-api-python-client==2.70.0
    # via -r requirements.in
google-auth==2.15.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via google-api-python-client
googleapis-common-protos==1.57.0
    # via google-api-core
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
jsonschema==4.17.3
    # via -r requirements.in
loguru==0.6.0
    # via -r requirements.in
mypy-extensions==0.4.3
    # via black
numpy==1.24.0
    # via -r requirements.in
packaging==23.0
    # via pytest
pathspec==0.10.3
    # via black
platformdirs==2.6.0
    # via black
pluggy==1.0.0
    # via pytest
protobuf==4.21.12
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.9
    # via httplib2
pyrsistent==0.19.3
    # via jsonschema
pytest==7.2.1
    # via -r requirements.in
python-dotenv==0.21.0
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
requests==2.28.1
    # via google-api-core
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
tomli==2.0.1
    # via
    #   black
    #   pytest
tqdm==4.64.1
    # via -r requirements.in
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.13
    # via requests
